# contracts/speech/speech.openapi.yaml
# Contract zone: verified by just contract-drift. Source: apps/jmcp-speechd + crates/jmcp-adapter-speech. DO NOT EDIT WITHOUT CONTRACT DRIFT.
openapi: 3.1.0
info:
  title: JMCP Talk Speech Runtime
  version: 1.0.0
  summary: Deterministic local speech daemon and sidecar-compatible speech API.
servers:
  - url: http://127.0.0.1:18879
paths:
  /health:
    get:
      operationId: speechHealth
      responses:
        "200":
          description: Speech runtime health.
  /metrics:
    get:
      operationId: speechMetrics
      responses:
        "200":
          description: Prometheus-compatible speech counters.
  /transcribe:
    post:
      operationId: transcribeAudio
      requestBody:
        required: true
        content:
          audio/wav:
            schema:
              type: string
              contentEncoding: base64
      responses:
        "200":
          description: Deterministic or sidecar ASR transcript.
  /synthesize:
    post:
      operationId: synthesizeSpeech
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text: { type: string }
                voice: { type: string }
                speed: { type: number }
      responses:
        "200":
          description: WAV audio bytes.
components:
  schemas:
    SpeechTraceEvent:
      type: object
      required:
        - turn_id
        - session_id
        - event
        - status
        - adapter
        - model
        - device
        - quantization
        - latency_ms
      properties:
        turn_id: { type: string }
        session_id: { type: string }
        event: { type: string }
        status: { type: string }
        adapter: { type: string }
        model: { type: string }
        device: { type: string }
        quantization: { type: string }
        latency_ms: { type: integer }
        error_class: { type: [string, "null"] }
        redacted_transcript: { type: [string, "null"] }
        audio_hash: { type: [string, "null"] }